requirements.txt
unidecode
geopandas
pyomo
tqdm
scikit-learn
statsmodels
plotly
dask
openpyxl